# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/bulk/rename": {
            "post": {
                "description": "Bulk Rename Files",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Bulk Rename Files",
                "parameters": [
                    {
                        "description": "Remote Configuration",
                        "name": "remote",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_kommunkod_restclone_pkg_rclone.BulkRenameFilesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/file/read": {
            "post": {
                "description": "Read File",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Read File",
                "parameters": [
                    {
                        "description": "Remote Configuration",
                        "name": "remote",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_kommunkod_restclone_pkg_rclone.ReadFileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kommunkod_restclone_pkg_rclone.ReadFileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/file/write": {
            "put": {
                "description": "Write File. You can attach an arbitrary number of files to the request. All have to be placed in the \"file\" field.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Write File",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Remote Configuration",
                        "name": "remote",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Path",
                        "name": "path",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Overwrite",
                        "name": "overwrite",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kommunkod_restclone_pkg_rclone.WriteFileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/files/moveBackupDir": {
            "post": {
                "description": "Move a file to a backup directory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Move Backup Dir",
                "parameters": [
                    {
                        "description": "Move Backup Dir Request",
                        "name": "moveBackupDirRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_kommunkod_restclone_pkg_rclone.MoveBackupDirRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Backup directory moved successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "config.UpdateRemoteOpt": {
            "type": "object",
            "properties": {
                "all": {
                    "description": "If set then ask all the questions, not just the post config questions",
                    "type": "boolean"
                },
                "continue": {
                    "description": "If set then supply state and result parameters to continue the process",
                    "type": "boolean"
                },
                "edit": {
                    "description": "If set then edit existing values",
                    "type": "boolean"
                },
                "noObscure": {
                    "description": "Treat all passwords as obscured",
                    "type": "boolean"
                },
                "nonInteractive": {
                    "description": "Don't interact with the user - return questions",
                    "type": "boolean"
                },
                "obscure": {
                    "description": "Treat all passwords as plain that need obscuring",
                    "type": "boolean"
                },
                "result": {
                    "description": "Result to return - used with Continue",
                    "type": "string"
                },
                "state": {
                    "description": "State to restart with - used with Continue",
                    "type": "string"
                }
            }
        },
...
        "github_com_kommunkod_restclone_pkg_rclone.BulkRenameFilesRequest": {
            "type": "object",
            "properties": {
                "nameMap": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "path": {
                    "type": "string"
                },
                "remote": {
                    "$ref": "#/definitions/github_com_kommunkod_restclone_pkg_rclone.RemoteConfiguration"
                }
            }
        },
...
        "github_com_kommunkod_restclone_pkg_rclone.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_kommunkod_restclone_pkg_rclone.FileItem": {
            "type": "object",
            "properties": {
                "mimeType": {
                    "type": "string"
                },
                "modTime": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_kommunkod_restclone_pkg_rclone.MoveBackupDirRequest": {
            "type": "object",
            "properties": {
                "destinationPath": {
                    "type": "string"
                },
                "destinationRemote": {
                    "$ref": "#/definitions/github_com_kommunkod_restclone_pkg_rclone.RemoteConfiguration"
                },
                "sourcePath": {
                    "type": "string"
                },
                "sourceRemote": {
                    "$ref": "#/definitions/github_com_kommunkod_restclone_pkg_rclone.RemoteConfiguration"
                }
            }
        },
...
        "github_com_kommunkod_restclone_pkg_rclone.ReadFileData": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kommunkod_restclone_pkg_rclone.FileItem"
                    }
                }
            }
        },
        "github_com_kommunkod_restclone_pkg_rclone.ReadFileRequest": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                },
                "remote": {
                    "$ref": "#/definitions/github_com_kommunkod_restclone_pkg_rclone.RemoteConfiguration"
                }
            }
        },
        "github_com_kommunkod_restclone_pkg_rclone.ReadFileResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kommunkod_restclone_pkg_rclone.ReadFileData"
                },
                "error": {
                    "$ref": "#/definitions/github_com_kommunkod_restclone_pkg_rclone.Error"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_kommunkod_restclone_pkg_rclone.RemoteConfiguration": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "options": {
                    "$ref": "#/definitions/config.UpdateRemoteOpt"
                },
                "parameters": {
                    "type": "object",
                    "additionalProperties": true
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "github_com_kommunkod_restclone_pkg_rclone.WriteFileData": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kommunkod_restclone_pkg_rclone.FileItem"
                    }
                }
            }
        },
        "github_com_kommunkod_restclone_pkg_rclone.WriteFileResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_kommunkod_restclone_pkg_rclone.WriteFileData"
                },
                "error": {
                    "$ref": "#/definitions/github_com_kommunkod_restclone_pkg_rclone.Error"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}